packages/db_mcp/pyproject.toml
[project]
name = "db-mcp"
version = "0.1.0"
description = "MCP Server for Roadmap Database"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.25.0",
    "psycopg2-binary>=2.9.11",
    "pgvector>=0.4.2",
    "google-genai>=1.59.0",
    "pydantic-settings>=2.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project.scripts]
db-mcp = "server:main"
